Milestone 2/requirements.txt
pandas
numpy
requests
matplotlib
# seaborn
# selenium
bs4
scikit-learn
SQLAlchemy
psycopg2-binary
PyYAML
docker 
kafka-python